requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
av==14.2.0
    # via py-feat
celluloid==0.2.0
    # via py-feat
certifi==2025.1.31
    # via requests
charset-normalizer==3.4.1
    # via requests
colorama==0.4.6
    # via tqdm
contourpy==1.3.0
    # via matplotlib
cycler==0.12.1
    # via matplotlib
easing-functions==1.0.4
    # via py-feat
filelock==3.18.0
    # via torch
fonttools==4.57.0
    # via matplotlib
fsspec==2025.3.2
    # via torch
h5py==3.13.0
    # via
    #   nltools
    #   py-feat
idna==3.10
    # via requests
imageio==2.37.0
    # via scikit-image
importlib-resources==6.5.2
    # via
    #   matplotlib
    #   nibabel
jinja2==3.1.6
    # via torch
joblib==1.4.2
    # via
    #   nilearn
    #   nltools
    #   py-feat
    #   scikit-learn
kiwisolver==1.4.7
    # via matplotlib
kornia==0.8.0
    # via py-feat
kornia-rs==0.1.8
    # via kornia
lazy-loader==0.4
    # via scikit-image
lxml==5.3.2
    # via nilearn
markupsafe==3.0.2
    # via jinja2
matplotlib==3.9.4
    # via
    #   celluloid
    #   nltools
    #   py-feat
    #   seaborn
mpmath==1.3.0
    # via sympy
networkx==3.2.1
    # via
    #   scikit-image
    #   torch
nibabel==5.3.2
    # via
    #   nilearn
    #   nltools
nilearn==0.11.1
    # via nltools
nltools==0.5.1
    # via py-feat
numexpr==2.8.4
    # via py-feat
numpy==1.23.5
    # via
    #   contourpy
    #   h5py
    #   imageio
    #   matplotlib
    #   nibabel
    #   nilearn
    #   nltools
    #   numexpr
    #   pandas
    #   py-feat
    #   pywavelets
    #   scikit-image
    #   scikit-learn
    #   scipy
    #   seaborn
    #   tifffile
    #   torchvision
    #   xgboost
packaging==24.2
    # via
    #   kornia
    #   lazy-loader
    #   matplotlib
    #   nibabel
    #   nilearn
    #   scikit-image
pandas==2.2.3
    # via
    #   nilearn
    #   nltools
    #   py-feat
    #   seaborn
pillow==11.1.0
    # via
    #   imageio
    #   matplotlib
    #   py-feat
    #   scikit-image
    #   torchvision
py-feat==0.6.2
    # via -r requirements.in
pynv==0.3
    # via nltools
pyparsing==3.2.3
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
pytz==2025.2
    # via pandas
pywavelets==1.6.0
    # via py-feat
requests==2.32.3
    # via
    #   nilearn
    #   pynv
scikit-image==0.24.0
    # via py-feat
scikit-learn==1.6.1
    # via
    #   nilearn
    #   nltools
    #   py-feat
scipy==1.13.1
    # via
    #   nilearn
    #   nltools
    #   scikit-image
    #   scikit-learn
    #   xgboost
seaborn==0.13.2
    # via
    #   nltools
    #   py-feat
six==1.17.0
    # via python-dateutil
sympy==1.13.1
    # via torch
threadpoolctl==3.6.0
    # via scikit-learn
tifffile==2024.8.30
    # via scikit-image
torch==2.6.0
    # via
    #   kornia
    #   torchvision
torchvision==0.21.0
    # via py-feat
tqdm==4.67.1
    # via py-feat
typing-extensions==4.13.1
    # via
    #   nibabel
    #   torch
tzdata==2025.2
    # via pandas
urllib3==2.3.0
    # via requests
xgboost==2.1.4
    # via py-feat
zipp==3.21.0
    # via importlib-resources
